Add function to compute displacement on device
git-svn-id: https://icl.cs.utk.edu/svn/magma/branches/sc_release@1010 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ control/magma_zf77.cpp
@@ -26,7 +26,7 @@
  * typedef comming from fortran.h file provided in $CUDADIR/src directory
  * it will probably change with future release of cublas when they will use 64bits address
  */
-typedef int devptr_t;
+typedef size_t devptr_t;
 
 #define PRECISION_z
